requirements.txt
fastapi
uvicorn
pillow
qrcode